requirements.txt
tensorflow
flask
gunicorn
numpy
pillow
opencv-python